+++ docs/examples/source/logging_ex.py
@@ -1,0 +1,29 @@
+import github3
+import logging
+
+# Set up a file to have all the logs written to
+file_handler = logging.FileHandler('github_script.log')
+
+# Send the logs to stderr as well
+stream_handler = logging.StreamHandler()
+
+# Format the log output and include the log level's name and the time it was
+# generated
+formatter = logging.Formatter('%(asctime)s %(levelname)s %(message)s')
+
+# Use that Formatter on both handlers
+file_handler.setFormatter(formatter)
+stream_handler.setFormatter(formatter)
+
+# Get the logger used by github3.py internally by referencing its name
+# directly
+logger = logging.getLogger('github3')
+# Add the handlers to it
+logger.addHandler(file_handler)
+logger.addHandler(stream_handler)
+# Set the level which determines what you see
+logger.setLevel(logging.DEBUG)
+
+# Make a library call and see the information posted
+r = github3.repository('sigmavirus24', 'github3.py')
+print('{0} - {0.html_url}'.format(r))